Fix some asset imports
