+++ ui/src/components/FavouritePostIndicator/index.ts
@@ -1,0 +1,1 @@
+export { FavouritePostIndicator } from "./FavouritePostIndicator";
