+++ ui/src/components/PostPreview/index.ts
@@ -1,0 +1,1 @@
+export { PostPreview } from "./PostPreview";
